*: refactor some functions for better reusability
